backend/requirements.txt
# FastAPI核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
alembic==1.12.1

# 数据验证和序列化
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
cryptography==41.0.7

# 工具库
python-dotenv==1.0.0
typing-extensions==4.8.0
click==8.1.7
typer==0.9.0

# HTTP客户端
httpx==0.27.0
requests==2.31.0

# AI库
openai>=1.97.0

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# 日志
loguru==0.7.2

# CORS支持
fastapi-cors==0.0.6

# 缓存
redis==5.0.1

# 异步任务
celery==5.3.4